requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib
requests
python-dotenv
asyncpg
sqlalchemy[asyncio]
aiosqlite